revert production app changes
