+++ src/main/notifications/index.ts
@@ -1,25 +1,25 @@
 // Copyright (c) 2016-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import {shell, Notification, ipcMain} from 'electron';
+import {app, shell, Notification, ipcMain} from 'electron';
 import isDev from 'electron-is-dev';
 import {getDoNotDisturb as getDarwinDoNotDisturb} from 'macos-notification-state';
 
 import MainWindow from 'app/mainWindow/mainWindow';
+import TabManager from 'app/tabs/tabManager';
 import WebContentsManager from 'app/views/webContentsManager';
-import {PLAY_SOUND, OPEN_NOTIFICATION_PREFERENCES} from 'common/communication';
+import {PLAY_SOUND, NOTIFICATION_CLICKED, BROWSER_HISTORY_PUSH, OPEN_NOTIFICATION_PREFERENCES} from 'common/communication';
+import Config from 'common/config';
 import {Logger} from 'common/log';
 import ServerManager from 'common/servers/serverManager';
 import viewManager from 'common/views/viewManager';
 import DeveloperMode from 'main/developerMode';
 import PermissionsManager from 'main/security/permissionsManager';
 
-import {dispatchMentionClick} from './dispatchMentionClick';
 import getLinuxDoNotDisturb from './dnd-linux';
 import getWindowsDoNotDisturb from './dnd-windows';
 import {DownloadNotification} from './Download';
 import {Mention} from './Mention';
-import {triggerNotificationFrameEffects} from './notificationFrameEffects';
 import {NewVersionNotification, UpgradeNotification} from './Upgrade';
 
 const log = new Logger('Notifications');
@@ -92,7 +92,16 @@ class NotificationManager {
             log.debug('notification click', server.id, mention.uId);
 
             this.allActiveNotifications?.delete(mention.uId);
-            dispatchMentionClick(view, webcontents, channelId, teamId, url);
+
+            // Show the window after navigation has finished to avoid the focus handler
+            // being called before the current channel has updated
+            const focus = () => {
+                MainWindow.show();
+                TabManager.switchToTab(view.id);
+                ipcMain.off(BROWSER_HISTORY_PUSH, focus);
+            };
+            ipcMain.on(BROWSER_HISTORY_PUSH, focus);
+            webcontents.send(NOTIFICATION_CLICKED, channelId, teamId, url);
         });
 
         mention.on('close', () => {
@@ -127,7 +136,7 @@ class NotificationManager {
                         if (notificationSound) {
                             MainWindow.sendToRenderer(PLAY_SOUND, notificationSound);
                         }
-                        triggerNotificationFrameEffects(true);
+                        flashFrame(true);
                         clearTimeout(timeout);
                         resolve({status: 'success'});
                     }
@@ -168,7 +177,7 @@ class NotificationManager {
         this.allActiveNotifications?.set(download.uId, download);
 
         download.on('show', () => {
-            triggerNotificationFrameEffects(true);
+            flashFrame(true);
         });
 
         download.on('click', () => {
@@ -256,6 +265,17 @@ export async function getDoNotDisturb() {
     }
 
     return false;
+}
+
+function flashFrame(flash: boolean) {
+    if (process.platform === 'linux' || process.platform === 'win32') {
+        if (Config.notifications.flashWindow) {
+            MainWindow.get()?.flashFrame(flash);
+        }
+    }
+    if (process.platform === 'darwin' && Config.notifications.bounceIcon && Config.notifications.bounceIconType) {
+        app.dock?.bounce(Config.notifications.bounceIconType);
+    }
 }
 
 const notificationManager = new NotificationManager();
